Remade the algorithm for camera sharing and added an ability for users to share their screen
(A brief explanation of the new algorithm) Now whenever the user joins the conference, they send the default mediaStreamId to other participants to be used for camera feed and microphone audio. Whenever the host starts screen sharing, the signaling server sends the mediaStream id of the screen media to the users and then adds the tracks and renegotiates the peer connections to initialize the remote screen feed.